feat(erro page): add error page and remove redirect
