+++ src/pages/index.tsx
@@ -8,11 +8,7 @@ import { pageview } from "@/utils/gtag";
 import { useCallback, useEffect } from "react";
 import { useRouter } from "next/router";
 import { NextSeo } from "next-seo";
-import { Inter } from "next/font/google";
 import { Support } from "@/components/Support";
-import { GetStaticPaths, GetStaticProps } from "next";
-import { PageSEO, SEO_MULTI_PAGE } from "../utils/SEO";
-const font = Inter({ subsets: ["latin"] });
 
 export default function Home() {
   const { isLogged, user } = useAuth();
@@ -37,28 +33,8 @@ export default function Home() {
         title="Crie Figurinhas WhatsApp Grátis em 3 Passos | Online e Divertido"
         description="Saiba como fazer figurinhas para WhatsApp de forma fácil e gratuita. Com nossa aplicação online, autentique-se, escolha a imagem e receba no WhatsApp. Diversão garantida! Começe agora."
       />
-      <div
-        className={`h-full lg:h-screen flex flex-col ${font.className} relative`}
-      >
-        <div className="flex items-center justify-center gap-x-6 bg-indigo-600 px-6 py-2.5 sm:px-3.5 ">
-          {/* sm:before:flex-1} */}
-          <p className="text-sm leading-6 text-white">
-            <span>
-              <strong className="font-semibold">Versão BETA</strong>
-              <svg
-                viewBox="0 0 2 2"
-                className="mx-2 inline h-0.5 w-0.5 fill-current"
-                aria-hidden="true"
-              >
-                <circle cx={1} cy={1} r={1} />
-              </svg>
-              Relaxe e aproveite a fase beta. Erros são apenas uma parte natural
-              da evolução. Nós cuidaremos deles.
-              {/* <span aria-hidden="true">&rarr;</span> */}
-            </span>
-          </p>
-        </div>
-        {/* <section className="shadow-sm">
+
+      {/* <section className="shadow-sm">
           <header className="container h-16 mx-auto flex items-center">
             <Github className="w-8 h-8" />
             {isLogged && (
@@ -69,36 +45,35 @@ export default function Home() {
           </header>
         </section> */}
 
-        <main className="w-full h-full max-w-7xl mx-auto text-center lg:text-left px-4 md:container mt-16">
-          <section className="flex flex-col gap-20 lg:gap-0 lg:flex-row items-center py-10">
-            <div className="flex flex-col flex-1 gap-4">
-              <h1 className="text-[56px] leading-tight font-bold text-zinc-800">
-                Fazer Figurinhas para WhatsApp Nunca Foi Tão Fácil!
-              </h1>
-              <h2 className="text-lg text-zinc-500">
-                Descubra a maneira mais simples, gratuita e online de criar suas
-                próprias figurinhas para WhatsApp. Com apenas três passos -
-                autenticar, selecionar imagem e receber no WhatsApp - você
-                estará gerando figurinhas engraçadas em um piscar de olhos.
-              </h2>
-            </div>
-            <UploadSticker />
-          </section>
-        </main>
+      <main className="w-full h-full max-w-7xl mx-auto text-center lg:text-left px-4 md:container mt-16">
+        <section className="flex flex-col gap-20 lg:gap-0 lg:flex-row items-center py-10">
+          <div className="flex flex-col flex-1 gap-4">
+            <h1 className="text-[56px] leading-tight font-bold text-zinc-800">
+              Fazer Figurinhas para WhatsApp Nunca Foi Tão Fácil!
+            </h1>
+            <h2 className="text-lg text-zinc-500">
+              Descubra a maneira mais simples, gratuita e online de criar suas
+              próprias figurinhas para WhatsApp. Com apenas três passos -
+              autenticar, selecionar imagem e receber no WhatsApp - você estará
+              gerando figurinhas engraçadas em um piscar de olhos.
+            </h2>
+          </div>
+          <UploadSticker />
+        </section>
+      </main>
 
-        <footer className="text-center font-bold text-xs p-4">
-          Criado Por{" "}
-          <Link
-            title="github do fundador do figurinhas Bruno"
-            href="https://github.com/brunooomelo"
-            target="_blank"
-            className="text-indigo-600"
-          >
-            Bruno Melo
-          </Link>
-        </footer>
-        <Support />
-      </div>
+      <footer className="text-center font-bold text-xs p-4">
+        Criado Por{" "}
+        <Link
+          title="github do fundador do figurinhas Bruno"
+          href="https://github.com/brunooomelo"
+          target="_blank"
+          className="text-indigo-600"
+        >
+          Bruno Melo
+        </Link>
+      </footer>
+      <Support />
       {!isLogged && <LoginForm />}
       {isLogged && !user?.isAuthenticated && !!user?.whatsapp && <PinForm />}
     </>
